+++ UIElementsUnturned.UIElementsLib/Core/UI/Holder/UIHolder.cs
@@ -1,0 +1,17 @@
+﻿using System.Collections.Generic;
+using UIElementsLib.Core.UI.Element;
+using UIElementsUnturned.UIElementsLib.Core.UI.Holder.Base;
+
+namespace UIElementsUnturned.UIElementsLib.Core.UI.Holder
+{
+    public sealed class UIHolder<TUIHolder> : UIHolderBase<TUIHolder> where TUIHolder : IUIElement
+    {
+        public UIHolder(IEnumerable<TUIHolder> items) : base(items)
+        {
+        }
+
+        public UIHolder() : base()
+        {
+        }
+    }
+}