+++ CHANGELOG.md
@@ -1,8 +1,18 @@
 | Versions: |
 | - |
+| [v1.0.1](#v101) |
 | [v1.0.0](#v100) |
 
 ---
+
+### v1.0.1:
+
+1. Updated all libraries to the latest versions
+2. Added PlayerUIElementsListenerContainer as well as IUIElementsContainer (special for easy usage of lib)
+3. Added new project with examples how to use lib in simple and fast way
+4. Updated documentation
+5. Added Default example
+6. Added Fast & simple example
 
 ### v1.0.0: